Completed task 3a and 3b
Defined all of the triangular sides, calculated the respective vectors (e.g. AB = b-a) using the classes defined previously, then took half of the dot product of two of these vectors to find the area. To determine the angles, i took the inverse cosign of the dot product of two vectors over the magnitude of each respective vector multiplied together and repeated this for each new set of vectors