Updates to previous 'Assignment3.py' (efficiency etc)
Updated .py file that imports MC module to conform to snake case (changed name from 'Assignment3.py to 'assignment3.py') and changed the way I used my module to calculate the MC volume of a 2D sphere as I was calculating points and using that as my radius, causing an error as points is an array of size (2, N-points) and I want the radius to be a float. fixed this by directly jumping to calculate my radius and having the points be generated in my module file as opposed to in this file. Also fixed my prints to be of a proper fstring format and limit the number of decimal places printed