Updates to MonteCarlo_class.py file
Fixed file naming system to fit with snake-case naming system (changed from MonteCarlo_class.py to montecarlo_class.py) and fixed all variable/parameter names within the file to follow this naming convention. Fixed the way I was generating random seeds to allow the user to fully define if the want to use a seed or not through use of an if else statement, meaning if seed is defined as anything but 'None' then, a random seed defined by the user will be used. Also began the proper implementation of using MPI to divide the sample number across processes, this is still a work in process but should be fully implemented soon. Also added in my licence to my code and updated most of the docstring descriptions within my code to be more informative to those using it.