Working integral function in class
Fixed issues I was having with the integral function. Redefined part of my statistics module to allow for filtering between integral and non integral values being used for statistics calculation, this allows me to produce a weighted mean and variance calculation for the integral as opposed to using the raw data (which falsely reports the variance as it gives small outputs). Updated some of the docstrings as well to make code more readable