'bones' of poisson class created
From task 1 of the assignment, I have implemented a function to define where the potential will be placed via use of boundary conditions, ensuring to encapsulate both the left and right and the top and bottom of the NxN grid. I also added a charge distribution function which will define the charge distribution across the grid in a similar fashion as was done for the potential. Finally I have implemented an overrelaxation function which updates a new value for the charge after overrelaxation has taken place, using the equation defined in the assignment 4 hand-out.